update shopByCategory to CategoryShowcase smart component
